chore: split up geo add test
