+++ packages/amplify-e2e-tests/src/__tests__/geo-add-e.test.ts
@@ -1,0 +1,64 @@
+/* eslint-disable spellcheck/spell-checker */
+/* eslint-disable import/no-extraneous-dependencies */
+
+import {
+  createNewProjectDir,
+  deleteProject,
+  deleteProjectDir,
+  initJSProjectWithProfile,
+  addAuthWithDefault,
+  addPlaceIndexWithDefault,
+  getProjectMeta,
+  amplifyPushWithoutCodegen,
+  getPlaceIndex,
+  generateRandomShortId,
+  getGeoJSConfiguration,
+} from '@aws-amplify/amplify-e2e-core';
+import { existsSync } from 'fs';
+import path from 'path';
+import { getAWSExports } from '../aws-exports/awsExports';
+
+describe('amplify geo add e', () => {
+  let projRoot: string;
+  beforeEach(async () => {
+    projRoot = await createNewProjectDir('geo-add-test');
+  });
+
+  afterEach(async () => {
+    const metaFilePath = path.join(projRoot, 'amplify', '#current-cloud-backend', 'amplify-meta.json');
+    if (existsSync(metaFilePath)) {
+      await deleteProject(projRoot);
+    }
+    deleteProjectDir(projRoot);
+  });
+
+  it('init a project with default auth config and add two place index resources with the second set to default', async () => {
+    const index1Id = `placeindex${generateRandomShortId()}`;
+    const index2Id = `placeindex${generateRandomShortId()}`;
+    await initJSProjectWithProfile(projRoot, {});
+    await addAuthWithDefault(projRoot);
+    await addPlaceIndexWithDefault(projRoot, { resourceName: index1Id, isFirstGeoResource: true });
+    await addPlaceIndexWithDefault(projRoot, { resourceName: index2Id, isAdditional: true });
+    await amplifyPushWithoutCodegen(projRoot);
+
+    // check amplify meta file
+    const meta = getProjectMeta(projRoot);
+    expect(meta.geo[index1Id].isDefault).toBe(false);
+    expect(meta.geo[index2Id].isDefault).toBe(true);
+    // check if resource is provisioned in cloud
+    const region = meta.geo[index1Id].output.Region;
+    const index1Name = meta.geo[index1Id].output.Name;
+    const index2Name = meta.geo[index2Id].output.Name;
+    const index1 = await getPlaceIndex(index1Name, region);
+    const index2 = await getPlaceIndex(index2Name, region);
+    expect(index1.IndexName).toBeDefined();
+    expect(index2.IndexName).toBeDefined();
+    // check aws export file
+    const awsExport: any = getAWSExports(projRoot).default;
+    expect(getGeoJSConfiguration(awsExport).search_indices.items).toContain(index1Name);
+    expect(getGeoJSConfiguration(awsExport).search_indices.items).toContain(index2Name);
+    expect(getGeoJSConfiguration(awsExport).search_indices.default).toEqual(index2Name);
+    expect(getGeoJSConfiguration(awsExport).region).toEqual(region);
+  });
+
+});
